The staged copy of HgiVulkanBlitCmds has incorrect source and destination offset
Source offset of a cpu-to-gpu-copy defines the offset of the source cpu memory,
and the destination offset of a cpu-to-gpu-copy defines the offset of the
destination gpu buffer.

The `memcpy`` from the cpu memory to the staging buffer should use the source offset
and the destination offset respectively. But the copy from the staging buffer
to the gpu buffer should use the destination offset to define `VkBufferCopy.srcOffset`.