Update HdPrman's implementation of PreviewSurface to match the updated spec
The preview surface spec was recently updated to state that there is no lighting
response for fully transparent materials. This change removes the refractionGain
connection between the adapter and pxrSurface nodes that appears to be
resulting in that additional lighting response.

Previously this connection was only added when in cutout mode. This meant that
when opacityThreshold was non-zero OR not authored we would get additional
lighting response for fully transparent materials.

This change means that HdPrman will match the new spec and fully transparent
materials (opacity = 0) will not have any lighting response regardless of if we are
in cutout or transparent mode.

(Internal change: 2345815)