Check prototypes before creating gprim instances
HdPrmanInstancer::Populate() expects to receive invalid geometry prototype ids when instancing analytic lights, so it checks to make sure it has a light shader id whenever it sees an invalid geometry prototype id and fails a verify if both are invalid.

We want to keep that behavior, but avoid hitting the failed verify when we expected to have valid geometry prototype ids. So we add a check in Gprim::Sync() to ensure we have valid ids before calling Populate() and issue a better (and non-fatal) warning if we do not.

(Internal change: 2345857)