Improvements to prim id rendering in Storm. One goal of this change is to move away from using "enableIdRender" in UsdImagingGLRenderParams and HdxRenderTaskParams to indicate prim id rendering, and instead move to requesting primId as a separate AOV. Currently, when one wants to do id rendering and sets enableIdRender to true, Storm will render prim ids to the default/color AOV. We instead want to use a primId AOV, which has an int32 format. This change adds support for rendering to an int32 format AOV.
This change also introduces the renderPassShaderKey, which is used to flexibly
generate a renderPassShader based on the desired AOVs. The AOVs currently
supported by this in Storm are color, primId and instanceId, and Neye.

There is still support for rendering primIds using the previous method i.e.
requesting it via "enableIdRender". This support will be deprecated in the
coming months, which will clean up some of the logic required to support both
xodepaths.

(Internal change: 2345969)