[hdSt] Fixes some test crash issues in the open source build.
I believe it's potentially problematic to declare an HdStRenderDelegate or
HdStResourceRegistrySharedPtr as static in the way some of these tests did.
The _HgiToResourceRegistryMap is a static singleton that includes a custom
deleter for HdStResourceRegistry. If _HgiToResourceRegistryMap is deleted first,
the custom deleter for HdStResourceRegistry might do a heap-use-after-free as
the deleter modifies a member of _HgiToResourceRegistryMap.

(Internal change: 2346349)