pcp: Improve recovery when TF_VERIFY fails
Pcp_ComposeExpressionVariables should always return a pointer to
a PcpExpressionVariables object whose lifetime is managed externally.
This is checked via a TF_VERIFY, but if that failed due to a bug
this function could return a pointer to a temporary instead. This
change avoids that by making that temporary static.

This allows client code to recover gracefully if the TF_VERIFY is
tripped. It also avoids a return-local-addr warning that is emitted
under gcc11.

(Internal change: 2346790)