Disable usdchecker for static builds and for nopy builds
Note that usdchecker as of right now depends on calling
UsdUtils.ComplianceChecker python module, which are not build in static
builds.

In a subsequent change, usdchecker will be updated to also work with the new
validation framework, which will remove this dependency on ComplianceChecker
and make it work with static builds.

(Internal change: 2346963)