2343722 added a new _didChangeSpecsAndChildrenInternal container to PcpCacheChanges for storing specs whose children also changed in order to limit the growth of the _didChangeSpecsInternal container resulting in a performance gain.  In the initial implementation, processing of this container was identical to that of _didChangeSpecsInternal.
The existing updateSpecStacks function contains code paths which may modify the state of the prim index cache during invocation.  Since the index cache is used for processing of _didChangeSpecsAndChildrenInternal, this could result in errors during iteration and processing if one of those branches were to ever be followed.

This change provides and more applicable name to this container and streamlines the processing of the items in _didChangePrimSpecsAndChildrenInternal to include only logic which ensures that prim stacks are updated.  Relevant property and target path caches are updated due to their parent paths having signficiantly changed due to addional / removal.

(Internal change: 2347031)