add ability to only load Hydra Scene Index Plugins for a given set of apps.
This adds a new supported field to HdSceneIndexPlugin type registrations:
"loadWithApps". Plugin types can filter down the set of apps that the
plugin will be auto-loaded for. At scene index creation time, an app name
can now be provided, and only plugins that either don't have an
loadWithApps entry or do contain the given app name in their loadWithApps
entry will be loaded. Not specifying an loadWithApps entry (or manually
specifying an empty list) both mean "should be loaded in all apps", which
means that all existing registered plugins will continue being loaded
normally.

Note that this new metadata only applies to plugin loading: it does not
control what happens once the plugin is loaded. If the plugin library is
loaded, the plugin will always be included in the created scene index. As a
result, the RegisterSceneIndexForRenderer method does _not_ have a new
appName parameter: if you're already at the point of calling the register
method, then the plugin has already been loaded, so the appName filtering
doesn't apply.

The render index also has to know about app names, because it is
responsible for creating scene indexes when scene index emulation is
enabled on the render index. The app name is provided as a new optional
argument to the HdRenderIndex constructor, which then passes that app name
through to the scene index plugin registry. Just like the plugin registry,
passing the empty string as the app name (the default) will load only the
non-app-specific scene index plugins.

This also makes one other enhancement to the scene index plugin APIs: it
allows HdSceneIndexPlugins to know what the render instance ID of the
render index is when creating scene indexes. Previously, this was exposed
to plugins that used the callback-style of scene index plugin registration,
but was not exposed to plugins that used the plugin-type-style
registration. A shim is provided so that existing scene index plugins don't
need to be modified.

(Internal change: 2348365)