[hdPrman] To have displacement work properly for UsdPreviewSurface's in Prman, we need to to have the primvar displacementbound:sphere set to 1 on the gprim. We previously accomplished this by adding a dummy material network node containing the primvar when we found a UsdPreviewSurface, and then had that primvar transferred to the gprim during the gprim sync. This approach led to warning spews however (as the dummy node was connected to a property that doesn't actually exist).
In this change, we take advantage of the HdsiMaterialPrimvarTransferSceneIndex,
which transfer primvars from a material to any prims using that material. We
set up scene index HdPrman_PreviewSurfacePrimvarsSceneIndex that happens before
the material primvar transfer, and adds the displacementbound:sphere primvar
to a material's primvars if that material has a UsdPreviewSurface node.

We maintain the old codepath for now to support USD versions in which the
HdPrman_MaterialPrimvarTransferSceneIndexPlugin is not available.

(Internal change: 2349710)