Add support for Scalar translate and scale xformOps
- Added translateX, translateY, translateZ, scaleX, scaleY and scaleZ
xformOps to correspond to translation along X, Y and Z and scale
along X, Y and Z respectively.
- Also added TypeTranslateX, TypeTranslateY, TypeTranslateZ, TypeScaleX,
TypeScaleY, TypeScaleZ types.
- Added APIs to UsdGeomXformable to allow to add the above mentioned
xforms to UsdGeomXform
- Updated testUsdGeomXformable to include test for adding / getting
these scalar translate and scale xformOps, and also make sure the
order in which these are set lead to a transformation which aligns
with the appropriate matrix multiplication math.

(Internal change: 2350257)