[usdImaging] Cards Cross Epsilon
Opposing axis faces of cards in cross geometry are coplanar. In GL, the backfaces are culled to avoid Z-fighting. But in renderman, the non-shaded backfaces can sometimes bleed through the opposing front-side face, producing a bit of a shadow in the crux of the cross. The bleed effect is nondeterministic, causing visible differences from render to render.

This change introduces a fixed, 2ε gap between opposing axis faces, which eliminates the erroneous and nondeterministic bleed.

(Internal change: 2351369)