build_usd.py: Create destination directory before running lipo
This fixes an error when performing a universal build on MacOS
that was introduced in change 2350749. The TBB build failed
to create the final universal .dylibs because the destination
"lib" directory did not exist, causing lipo to error out.
Prior to change 2350749, the zlib build would always run
before TBB and create that directory.

(Internal change: 2351618)