HdEngine: Adding overload of Execute taking task paths.
This is motivated by my work to turn the task controller into a scene index. The scene index would only give a client the paths to tasks. So clients would benefit from this HdEngine API.

That is, clients will not need to know about the HdRenderIndex::GetTask API.

(Internal change: 2351943)