Fix an incorrect memory deallocation.
TsRegressionPreventer::_WorkingKnotState attempted to use TsKnot to store
degenerate Ts_KnotData, with no typed parameters.  This didn't work when it came
time to delete the TsKnot.

Instead, store the time params in a naked Ts_KnotData.  In cases where a full
proposed knot was provided, store that also.  The duplication is minimal.

(Internal change: 2352157)