Add new opacityMode input to Preview Surface to allow for fully transparent materials to either have a specular response or not.
We recently updated the Preview Surface spec to follow the current Storm behavior
of fully transparent materials to not have a specular response as that is the desired
behavior when using Texture Cards. However, due to feedback from external users
this change adds an additional 'opacityMode' input to allow fully transparent
materials to either have a specular response 'transparent' or not 'presence'.
Note that Storm is unchanged so the only the 'presence' mode is currently supported in Storm, however both modes are supported in HdPrman.

(Internal change: 2352266)