Add MaterialXConfigAPI and dataSourceAttributeTypeName to help support USD files that were written with v1.38 MaterialX nodes but run with MaterialX v1.39
The MaterialXConfigAPI allows for the authored version to be recored in the Material Network.
This is then used in HdMtlx to record the appropriate version in the generated MaterialX
document. We can then make use of MaterialX's document upgrade function to perform
any necessary upgrades.

the dataSourceAttributeTypeName makes it possible to rely on the type names from the USD
file instead of turning to MaterialX's nodedefs. This is important to handle the case where
there have been changes to a node's input names between versions. Without this we would
not be able to determine the appropriate type. This can be re-used to help distinguish
between a color from a vector for example since the corresponding VtValue for both cases
would be a GfVector3.

(based on github pull request #3157)

Fixes #3157

(Internal change: 2352302)