[hdPrman] Set fixedSampleCount only on eligible lights
In 2351859, we added the HD_PRMAN_ALL_LIGHTS_FIXED_SAMPLE_COUNT environment variable to allow overriding fixedSampleCount on all lights hdPrman sends to RenderMan. But that parameter is only valid on a subset of lights. To avoid warnings when the parameter is set on ineligible lights, we now check whether the light supports the parameter before setting it whenever the environment variable is set.

(Internal change: 2352652)