Hydra emulation: making tasks also go through emulation. More precisely, the emulated scene index driven by legacy scene delegates.
This fixes the crash reported in https://github.com/PixarAnimationStudios/OpenUSD/issues/3471 caused by change 2351061.

The problem was that the HdxTaskController added a task without going through emulation. But when we received a removed notice (in this case UsdImagingStageSceneIndex removing /), we removed the task without re-adding it. That is, the merging scene index is removing and re-adding prims. But because the task was not properly emulated, it was never re-added when the merging scene index send the prims added notice.

Fixes #3471

(Internal change: 2353415)